Create gyrowithwifi.ino
This code contains the wifi part of the Charpy robot from UROCK 2022. This code enables user to be able to connect a internet connection and enables the user to read the focused movements that needed to be monitored by connecting it to an ip address which brings the user to the html file.